add http/https transport support

+++ lib/JSON/RPC2/AnyEvent/Client.pm
@@ -5,9 +5,10 @@ use warnings;
 
 use utf8;
 use AnyEvent::Handle;
+use AnyEvent::HTTP;
 use JSON::RPC2::Client;
 
-our $VERSION = "0.01";
+our $VERSION = "0.02";
 
 our $AUTOLOAD;  # it's a package global
 
@@ -20,7 +21,7 @@ sub new {
       call => 'call',
       @_,
       remappable => {},
-      cb => [],
+      cb => {},
    }, $class;
    $self->{call} = 'call'       if 'listed' eq $self->{call};
    $self->{call} = 'call_named' if 'named'  eq $self->{call};
@@ -31,20 +32,26 @@ sub new {
           $self->{remappable}->{ $_ } = '__'.$_ 
       }
    }
-   $self->__connect;
+   if ( $self->{url} ) {
+      $self->{request_fn} = \&JSON::RPC2::AnyEvent::Client::__request_http;
+   } else {
+      $self->__connect_tcp;
+      $self->{request_fn} = \&JSON::RPC2::AnyEvent::Client::__request_tcp;
+   }
    $self;
 }
 
-sub __connect {
+sub __connect_tcp {
    my $self = shift;
+   return if $self->{http};
    $self->{handle} = new AnyEvent::Handle
       connect  => [ $self->{host}, $self->{port} ],
       on_error => sub {
-         $_->("HTTP/1.0 500 $!") for @{$self->{cb}};
+         $self->__error($!);
          $self->{handle}->destroy; # explicitly destroy handle
       },
       on_eof   => sub {
-         $_->("CONNECTION CLOSED $!") for @{$self->{cb}};
+         $self->__error("CONNECTION CLOSED $!");
          $self->{handle}->destroy; # explicitly destroy handle
       };
 }
@@ -67,6 +74,14 @@ sub __service {
    $self;
 }
 
+sub __error {
+   my ( $self, $error ) = @_;
+   foreach my $call_id ( keys %{$self->{cb}} ) {
+      my $cb = delete $self->{cb}->{$call_id};
+      $cb->( $error );
+   }
+}
+
 sub AUTOLOAD {
    my $self = shift;
 
@@ -80,30 +95,57 @@ sub AUTOLOAD {
 
    my $cb = pop;
 
-   push @{$self->{cb}}, $cb;
-
    $method = $self->{service} ? $self->{service} . '.' . $method : $method;
 
    my $call = $self->{call};
 
    my ( $json_request, $call_id ) = $self->{client}->$call( $method, @_ );
 
+   $self->{cb}->{$call_id} = $cb;
+
+   $self->{request_fn}->( $self, $json_request );
+
+   return $call_id;
+}
+
+sub __request_tcp {
+   my ( $self, $json_request ) = @_;
+
    $self->{handle}->push_write( $json_request );
 
    $self->{handle}->push_read( json => sub{
-       my ( $handle, $hash ) = @_;
-       my ( $failed, $result, $error, $call_id ) = $self->{client}->response( $hash );
-       my $cb = shift @{$self->{cb}};
-       $cb->( $failed, $result, $error, $call_id );
+      my ( $handle, $hash ) = @_;
+      my ( $failed, $result, $error, $call_id ) = $self->{client}->response( $hash );
+      return $self->__error( $failed ) if $failed;
+      my $cb = delete $self->{cb}->{$call_id};
+      $cb->( $failed, $result, $error );
    } );
-   
-   return $call_id;
+}
+
+sub __request_http {
+   my ( $self, $json_request ) = @_;
+
+   http_post $self->{url}, $json_request, sub {
+      my ( $resp, $hdr ) = @_;
+
+      unless ( $hdr->{Status} =~ /^2/ ) {
+         return $self->__error( "$hdr->{Status} $hdr->{Reason}" );
+      }
+
+      my ( $failed, $result, $error, $call_id ) =
+         $self->{client}->response( $resp );
+
+      return $self->__error( $failed ) if $failed;
+
+      my $cb = delete $self->{cb}->{$call_id};
+      $cb->( $failed, $result, $error );
+   };
 }
 
 # This DESTROY-pattern originates from AnyEvent::Handle code.
 sub DESTROY {
    my ($self) = @_;
-   $self->{handle}->destroy();
+   $self->{handle}->destroy() if $self->{handle};
 }
 
 sub __destroy {
@@ -124,33 +166,32 @@ __END__
 
 =head1 NAME
 
-JSON::RPC2::AnyEvent::Client - Nonblocking JSON RPC2 client with method mapping.
+JSON::RPC2::AnyEvent::Client - Asynchronous nonblocking JSON RPC2 client with method mapping
 
 =head1 SYNOPSIS
 
     use JSON::RPC2::AnyEvent::Client;
 
-    # create connection
+    # create tcp connection
     my $rpc = JSON::RPC2::AnyEvent::Client->new(
         host    => "127.0.0.1",
         port    => 5555,
     );
 
     # call
-    $rpc->rpcfn( 1, 'two', 3, sub{
+    $rpc->sum( 1, 2, sub{
         my ( $failed, $result, $error ) = @_;
-        print Dumper $result if ! $failed && ! $error;
+        print $result unless $failed || $error;
     })
 
-    # call remote function with some configure
+    # call remote function with simple configure
     $rpc->service('agent')->listed()->remote_function( 'param1', 'param2', sub{
         my ( $failed, $result, $error ) = @_;
     })
 
-    # more arguments desctibed below
+    # some more constructor arguments
     my $rpc = JSON::RPC2::AnyEvent::Client->new(
-        host    => "127.0.0.1",
-        port    => 5555,
+        url     => "https://$host:$port/api", # http/https transport
         service => 'agent',
         call    => 'listed' || 'named',
         service => '_service',  # rename any this module methods
@@ -162,14 +203,24 @@ JSON::RPC2::AnyEvent::Client - Nonblocking JSON RPC2 client with method mapping.
 
 =head1 DESCRIPTION
 
-JSON::RPC2::AnyEvent::Client is JSON RPC2 client, currently with
-tcp transport, handled by L<AnyEvent::Handle>. Based on
-JSON RPC2 implementation L<JSON::RPC2::Client>. Remote
-functions is mapped to local client object methods. For example
-remote function fn(...) is called with local client as $c->fn(...,cb).
-Params of function is parms of remote functions with additional one
-at the end of param list. Additional last param is result handler
-soubroutine.
+JSON::RPC2::AnyEvent::Client is JSON RPC2 client, with
+tcp/http/https transport. Remote functions is mapped to local
+client object methods. For example remote function fn(...) is
+called as $c->fn(...,cb). Params of function is params of remote
+functions with additional one at the end of param list.
+Additional last param is result handler soubroutine.
+
+Implementation is based on JSON RPC2 implementation
+L<JSON::RPC2::Client>. Transport implementation is based
+on L<AnyEvent::Handle> for tcp, and on L<AnyEvent::HTTP>
+for http/https.
+
+The 'tcp' implementation use persistent connection, that make
+tcp connection at object creation and use it all object life time.
+The http/https persistence is AnyEvent::HTTP implementation
+dependent and currently it is not persistent for idempotent
+requsests (JSON RPC2 need POST requset). See description of
+'persistent' and 'keepalive' params of L<AnyEvent::HTTP>.
 
 =head1 METHODS
 
@@ -177,17 +228,25 @@ soubroutine.
 
 =item $rpc = B<new> JSON::RPC2::AnyEvent::Client host=>'example.com', ...
 
-The constructor supports these arguments (all as C<< key => value >> pairs).
+The constructor supports arguments as C<< key => value >> pairs.
 
 =over 4
 
 =item host => 'example.com'
 
-The hostname or ip address. The special value "unix/" used to connect to unix domain socket.
+The hostname or ip address. This enable tcp transport.
+The special value "unix/" used to connect to unix domain
+socket. Current version support unix domain socket only
+for 'tcp' transport.
 
 =item port => 5555
 
-The tcp port number or unix domain socket path.
+The tcp port number or unix domain socket path. Used togather
+with 'host' param.
+
+=item url => "https://$host:$port/api/rpc"
+
+The url of requst. This enables http/https transport.
 
 =item service => 'agent'
 
@@ -222,30 +281,41 @@ RPC named call type will be used.
 =item B<any other name> ( $param1, $param2, ..., $cb )
 
 Any method name will called via RPC on remote server. 
-Last param must be event handler cb(). 
-There is params of cb ( $fail, $result, $error );
-Where $result is server responce, valid only when there
-is no fail or error. The $error is remote server RPC error responce,
-is valid only when no fail. The $fail is transport error.
+Last param must be result handler callback cb(). 
 
 =back
 
+=head1 RESULT HANDLER CALLBACK
+
+The result callback handler is a soubroutine that called
+when rpc function is called and result is arrived or
+an error occured. There three param of callback is
+( $fail, $result, $error );
+
+The $fail is transport error. It is string that contain
+description of communication or data decoding error.
+
+The $result is server responce, valid only when there
+is no fail or error.
+
+The $error is described in rpc protocol standart remote
+server error responce. It is valid only when no fail.
 
 =head1 DEPENDENCIES
 
 =over 8
 
-=item L<AnyEvent::Socket>;
-
 =item L<AnyEvent::Handle>;
 
+=item L<AnyEvent::HTTP>;
+
 =item L<JSON::RPC2::Client>;
+
+=item L<JSON::XS>
 
 =back
 
 =head1 LICENSE
-
-Copyright (C) Serguei Okladnikov.
 
 This library is free software; you can redistribute it and/or modify
 it under the same terms as Perl itself.
